# docs/swagger.yaml
definitions:
...
  internal_product-service_transport_http_controller.SearchProductsResponse:
    properties:
      products:
        items:
          $ref: '#/definitions/marketplace_internal_product-service_domain.Product'
        type: array
    type: object
  marketplace_internal_product-service_domain.Product:
    properties:
      attributes:
        additionalProperties: true
        type: object
      category_id:
        type: string
      category_name:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      images:
        items:
          $ref: '#/definitions/marketplace_internal_product-service_domain.ProductImage'
        type: array
      name:
        type: string
      price:
        type: number
      seller_id:
        type: string
      status:
        type: string
      stock_count:
        type: integer
      updated_at:
        type: string
    type: object
  marketplace_internal_product-service_domain.ProductImage:
    properties:
      id:
        type: string
      image_url:
        type: string
      is_main:
        type: boolean
      sort_order:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /products/search:
    get:
      consumes:
      - application/json
      description: Search for products using vector search
      parameters:
      - description: Limit results
        in: query
        name: limit
        type: integer
      - description: Search query
        in: query
        name: query
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_product-service_transport_http_controller.SearchProductsResponse'
      summary: Search products
      tags:
      - products
...
swagger: "2.0"